pyproject.toml
[project]
name = "clashnodeeditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruamel-yaml>=0.19.1",
]
